backend/bot/requirements.txt
discord.py
supabase
python-dotenv
httpx
PyPDF2
